// src/api/client/models/CommitWithAnalysis.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { Commit } from './Commit';
import type { CoverageAnalysis } from './CoverageAnalysis';

export type CommitWithAnalysis = {
  commit: Commit;
  coverage?: CoverageAnalysis;
};